refactor: Update timestamp handling in ETL job for float precision
- Changed the timestamp handling in the ETL job to accept float Unix timestamps directly, improving performance and precision.
- Updated related tests to ensure compatibility with float timestamps, including validation for float precision in the output.
- Adjusted comments to reflect the new timestamp data type and its implications for ClickHouse integration.

+++ etl_job.py
@@ -449,10 +449,12 @@ class EtlJob:
 
                     for value_pair in series.get("values", []):
                         try:
-                            # ClickHouse DateTime requires integer Unix timestamp
-                            # Prometheus API returns timestamp as integer
-                            # (whole seconds) in JSON
-                            ts = int(value_pair[0])
+                            # Prometheus API returns timestamp as float
+                            # (seconds.fraction)
+                            # ClickHouse DateTime64(6) accepts float Unix timestamp
+                            # directly. No conversion needed - maximum performance
+                            # and precision
+                            ts = float(value_pair[0])
                             value = float(value_pair[1])
                         except (TypeError, ValueError, IndexError) as exc:
                             logger.warning(
